node[Plot.selection]
