+++ src/marks/brush.js
@@ -1,6 +1,7 @@
 import {brush as brusher, brushX as brusherX, brushY as brusherY, create, select} from "d3";
 import {identity, maybeTuple} from "../options.js";
-import {Mark, selectionEquals} from "../plot.js";
+import {Mark} from "../plot.js";
+import {selection, selectionEquals} from "../selection.js";
 import {applyDirectStyles, applyIndirectStyles} from "../style.js";
 
 const defaults = {
@@ -32,33 +33,33 @@ export class Brush extends Mark {
         .call((X && Y ? brusher : X ? brusherX : brusherY)()
           .extent([[marginLeft, marginTop], [width - marginRight, height - marginBottom]])
           .on("start brush end", function(event) {
-            const {type, selection} = event;
+            const {type, selection: extent} = event;
             // For faceting, when starting a brush in a new facet, clear the
             // brush and selection on the old facet. In the future, we might
             // allow independent brushes across facets by disabling this?
             if (type === "start" && brush.currentElement !== this) {
               if (brush.currentElement !== null) {
                 select(brush.currentElement).call(event.target.clear, event);
-                brush.currentElement.selection = null;
+                brush.currentElement[selection] = null;
               }
               brush.currentElement = this;
             }
             let S = null;
-            if (selection) {
+            if (extent) {
               S = index;
               if (X) {
-                let [x0, x1] = Y ? [selection[0][0], selection[1][0]] : selection;
+                let [x0, x1] = Y ? [extent[0][0], extent[1][0]] : extent;
                 if (x.bandwidth) x0 -= x.bandwidth();
                 S = S.filter(i => x0 <= X[i] && X[i] <= x1);
               }
               if (Y) {
-                let [y0, y1] = X ? [selection[0][1], selection[1][1]] : selection;
+                let [y0, y1] = X ? [extent[0][1], extent[1][1]] : extent;
                 if (y.bandwidth) y0 -= y.bandwidth();
                 S = S.filter(i => y0 <= Y[i] && Y[i] <= y1);
               }
             }
-            if (!selectionEquals(this.selection, S)) {
-              this.selection = S;
+            if (!selectionEquals(this[selection], S)) {
+              this[selection] = S;
               this.dispatchEvent(new Event("input", {bubbles: true}));
             }
           }))
@@ -67,7 +68,7 @@ export class Brush extends Mark {
           .call(applyIndirectStyles, options)
           .call(applyDirectStyles, options))
       .node();
-    g.selection = null;
+    g[selection] = null;
     return g;
   }
 }
